+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 Unreleased in the current development version (target v0.22.0):
 
 AQUA core complete list:
+- New `filter_key` option to filter NetCDF files before loading them based on year (#2543)
 - Removed plt.close() in plotting functions (#2538)
 - Documentation aligned to the new repository structure (#2525)
 - Complete workflow for pypi version to be run once per week (#2527)